# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "CustomizerSchema": {
        "properties": {
          "groups": {
            "items": {
              "type": "string"
            },
            "title": "Groups",
            "type": "array"
          },
          "parameters": {
            "items": {
              "$ref": "#/components/schemas/Parameter"
            },
            "title": "Parameters",
            "type": "array"
          },
          "source_sha256": {
            "default": "",
            "title": "Source Sha256",
            "type": "string"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "title": "CustomizerSchema",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ModelVersion": {
        "properties": {
          "author": {
            "title": "Author",
            "type": "string"
          },
          "commit": {
            "title": "Commit",
            "type": "string"
          },
          "current": {
            "default": false,
            "title": "Current",
            "type": "boolean"
          },
          "date": {
            "format": "date-time",
            "title": "Date",
            "type": "string"
          },
          "files": {
            "items": {
              "$ref": "#/components/schemas/VersionFile"
            },
            "title": "Files",
            "type": "array"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "short": {
            "title": "Short",
            "type": "string"
          }
        },
        "required": [
          "commit",
          "short",
          "author",
          "date",
          "message"
        ],
        "title": "ModelVersion",
        "type": "object"
      },
...
      "Option": {
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          },
          "value": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "integer"
              },
              {
                "type": "number"
              },
              {
                "type": "string"
              }
            ],
            "title": "Value"
          }
        },
        "required": [
          "name",
          "value"
        ],
        "title": "Option",
        "type": "object"
      },
...
      "Parameter": {
        "properties": {
          "accept": {
            "items": {
              "type": "string"
            },
            "title": "Accept",
            "type": "array"
          },
          "caption": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Caption"
          },
          "group": {
            "default": "",
            "title": "Group",
            "type": "string"
          },
          "initial": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "integer"
              },
              {
                "type": "number"
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Initial"
          },
          "max": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max"
          },
          "max_length": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Length"
          },
          "min": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Min"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "options": {
            "items": {
              "$ref": "#/components/schemas/Option"
            },
            "title": "Options",
            "type": "array"
          },
          "samples": {
            "items": {
              "type": "string"
            },
            "title": "Samples",
            "type": "array"
          },
          "step": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Step"
          },
          "type": {
            "enum": [
              "number",
              "integer",
              "string",
              "boolean",
              "select",
              "color",
              "font",
              "slider",
              "file"
            ],
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "name",
          "type"
        ],
        "title": "Parameter",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "VersionDiff": {
        "properties": {
          "base": {
            "title": "Base",
            "type": "string"
          },
          "files": {
            "items": {
              "$ref": "#/components/schemas/VersionFile"
            },
            "title": "Files",
            "type": "array"
          },
          "head": {
            "title": "Head",
            "type": "string"
          },
          "patch": {
            "title": "Patch",
            "type": "string"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          }
        },
        "required": [
          "slug",
          "base",
          "head",
          "patch"
        ],
        "title": "VersionDiff",
        "type": "object"
      },
      "VersionFile": {
        "description": "``A``/``M``/``D`` plus the path, relative to the model's own directory.",
        "properties": {
          "path": {
            "title": "Path",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status",
          "path"
        ],
        "title": "VersionFile",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Self-hosted OpenSCAD customizer for Bambuddy.",
    "title": "ScadBuddy",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/models/{slug}/versions": {
      "get": {
        "operationId": "list_versions_api_v1_models__slug__versions_get",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "maxLength": 108,
              "pattern": "^(builtin:)?[a-z0-9][a-z0-9-]*$",
              "title": "Slug",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ModelVersion"
                  },
                  "title": "Response List Versions Api V1 Models  Slug  Versions Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "A model's revision history",
        "tags": [
          "versions"
        ]
      }
    },
    "/api/v1/models/{slug}/versions/{commit}/diff": {
      "get": {
        "operationId": "get_version_diff_api_v1_models__slug__versions__commit__diff_get",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "maxLength": 108,
              "pattern": "^(builtin:)?[a-z0-9][a-z0-9-]*$",
              "title": "Slug",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "commit",
            "required": true,
            "schema": {
              "pattern": "^[0-9a-f]{7,40}$",
              "title": "Commit",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "base",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^[0-9a-f]{7,40}$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Base"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VersionDiff"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Diff a revision against another (its parent by default)",
        "tags": [
          "versions"
        ]
      }
    },
    "/api/v1/models/{slug}/versions/{commit}/restore": {
      "post": {
        "operationId": "restore_version_api_v1_models__slug__versions__commit__restore_post",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "maxLength": 108,
              "pattern": "^(builtin:)?[a-z0-9][a-z0-9-]*$",
              "title": "Slug",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "commit",
            "required": true,
            "schema": {
              "pattern": "^[0-9a-f]{7,40}$",
              "title": "Commit",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelVersion"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Restore a revision as a new commit",
        "tags": [
          "versions"
        ]
      }
    },
    "/api/v1/models/{slug}/versions/{commit}/schema": {
      "get": {
        "operationId": "get_version_schema_api_v1_models__slug__versions__commit__schema_get",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "maxLength": 108,
              "pattern": "^(builtin:)?[a-z0-9][a-z0-9-]*$",
              "title": "Slug",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "commit",
            "required": true,
            "schema": {
              "pattern": "^[0-9a-f]{7,40}$",
              "title": "Commit",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomizerSchema"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "A revision's customizer schema",
        "tags": [
          "versions"
        ]
      }
    },
    "/api/v1/models/{slug}/versions/{commit}/source": {
      "get": {
        "operationId": "get_version_source_api_v1_models__slug__versions__commit__source_get",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "maxLength": 108,
              "pattern": "^(builtin:)?[a-z0-9][a-z0-9-]*$",
              "title": "Slug",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "commit",
            "required": true,
            "schema": {
              "pattern": "^[0-9a-f]{7,40}$",
              "title": "Commit",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/plain": {}
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "A revision's OpenSCAD source",
        "tags": [
          "versions"
        ]
      }
    },
...
  }
}